unify queue structures
