--- node/src/ckd/mod.rs
@@ -1,3 +1,0 @@
-mod debug;
-mod metrics;
-pub mod queue;
