+++ node/src/coordinator.rs
@@ -1,4 +1,3 @@
-use crate::storage::CKDRequestStorage;
 use crate::config::{ConfigFile, MpcConfig, ParticipantsConfig, SecretsConfig};
 use crate::db::{DBCol, SecretDB, EPOCH_ID_KEY};
 use crate::indexer::handler::ChainBlockUpdate;
@@ -21,6 +20,7 @@ use crate::providers::ckd::CKDProvider;
 use crate::providers::eddsa::{EddsaSignatureProvider, EddsaTaskId};
 use crate::providers::{EcdsaSignatureProvider, EcdsaTaskId};
 use crate::runtime::AsyncDroppableRuntime;
+use crate::storage::CKDRequestStorage;
 use crate::storage::SignRequestStorage;
 use crate::tracking::{self};
 use crate::web::DebugRequest;
