+++ node/src/indexer/fake.rs
@@ -6,8 +6,8 @@ use crate::config::ParticipantsConfig;
 use crate::indexer::handler::CKDRequestFromChain;
 use crate::indexer::types::ChainCKDRespondArgs;
 use crate::requests::recent_blocks_tracker::tests::TestBlockMaker;
-use crate::types::SignatureId;
 use crate::tracking::{AutoAbortTask, AutoAbortTaskCollection};
+use crate::types::SignatureId;
 use anyhow::Context;
 use mpc_contract::config::Config;
 use mpc_contract::primitives::{
