+++ node/src/lib.rs
@@ -20,7 +20,6 @@ mod assets;
 #[cfg(test)]
 mod async_testing;
 mod background;
-pub mod ckd;
 pub mod cli;
 mod config;
 mod coordinator;
@@ -39,7 +38,6 @@ mod providers;
 pub mod requests;
 mod runtime;
 mod storage;
-pub mod signing;
 pub mod tracing;
 mod tracking;
 pub mod types;
