+++ node/src/providers/ckd/mod.rs
@@ -9,10 +9,10 @@ use mpc_contract::{
 use threshold_signatures::ecdsa::KeygenOutput;
 
 use crate::{
-    storage::CKDRequestStorage,
     config::{ConfigFile, MpcConfig},
     network::{MeshNetworkClient, NetworkTaskChannel},
     primitives::MpcTaskId,
+    storage::CKDRequestStorage,
     types::CKDId,
 };
