+++ node/src/requests/mod.rs
@@ -1,2 +1,4 @@
+mod debug;
+mod metrics;
 pub mod queue;
 pub mod recent_blocks_tracker;
