+++ node/src/types.rs
@@ -1,18 +1,30 @@
-use mpc_contract::primitives::{domain::DomainId, signature::{Payload, Tweak}};
+use std::fmt;
+
+use mpc_contract::primitives::{
+    domain::DomainId,
+    signature::{Payload, Tweak},
+};
 use near_indexer_primitives::CryptoHash;
 use near_sdk::{AccountId, PublicKey};
 use serde::{Deserialize, Serialize};
+
+pub enum RequestType {
+    Signature,
+    CKD,
+}
 
 pub type RequestId = CryptoHash;
 
 /// The trait that defines common functionality of MPC requests:
 /// currently CKD and signatures
-pub trait Requests {
-    fn get_id() -> RequestId;
-    fn get_receipt_id() -> CryptoHash;
-    fn get_entropy() -> [u8; 32];
-    fn get_timestamp_nanosec() -> u64;
-    fn get_domain_id() -> DomainId;
+pub trait Request {
+    fn get_id(&self) -> RequestId;
+    fn get_receipt_id(&self) -> CryptoHash;
+    fn get_entropy(&self) -> [u8; 32];
+    fn get_timestamp_nanosec(&self) -> u64;
+    fn get_domain_id(&self) -> DomainId;
+    // fn get_name() -> String;
+    fn get_type() -> RequestType;
 }
 
 pub type CKDId = CryptoHash;
@@ -30,7 +42,6 @@ pub struct CKDRequest {
     pub domain_id: DomainId,
 }
 
-
 pub type SignatureId = CryptoHash;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -44,4 +55,65 @@ pub struct SignatureRequest {
     pub entropy: [u8; 32],
     pub timestamp_nanosec: u64,
     pub domain: DomainId,
+}
+
+impl fmt::Display for RequestType {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            RequestType::Signature => write!(f, "signature"),
+            RequestType::CKD => write!(f, "ckd"),
+        }
+    }
+}
+
+impl Request for CKDRequest {
+    fn get_id(&self) -> RequestId {
+        self.id
+    }
+
+    fn get_receipt_id(&self) -> CryptoHash {
+        self.receipt_id
+    }
+
+    fn get_entropy(&self) -> [u8; 32] {
+        self.entropy
+    }
+
+    fn get_timestamp_nanosec(&self) -> u64 {
+        self.timestamp_nanosec
+    }
+
+    fn get_domain_id(&self) -> DomainId {
+        self.domain_id
+    }
+
+    fn get_type() -> RequestType {
+        RequestType::CKD
+    }
+}
+
+impl Request for SignatureRequest {
+    fn get_id(&self) -> RequestId {
+        self.id
+    }
+
+    fn get_receipt_id(&self) -> CryptoHash {
+        self.receipt_id
+    }
+
+    fn get_entropy(&self) -> [u8; 32] {
+        self.entropy
+    }
+
+    fn get_timestamp_nanosec(&self) -> u64 {
+        self.timestamp_nanosec
+    }
+
+    fn get_domain_id(&self) -> DomainId {
+        self.domain
+    }
+
+    fn get_type() -> RequestType {
+        RequestType::Signature
+    }
 }
